Rename "migrations" directory to "south_migrations"
This allows for the "migrations" directory to be used for
django-migrations (as of Django 1.7 and later). Older versions of
Django will still get the South migrations by default as long as South
1.0 or later is used, (and even with older South, the
SOUTH_MIGRATION_MODULES setting can be used to point to the
"south_migrations" directory).

This change is recommended in the release notes for South 1.0 here:

http://south.readthedocs.org/en/latest/releasenotes/1.0.html#library-migration-path